🤖 Refactor(client): background component CanvasFlow 폴더로 이동
